Refactor TestingHdfsEnvironment
Moving definition for class TestingHdfsEnvironment to a common place.
