+++ presto-hive/src/test/java/com/facebook/presto/hive/TestingHdfsEnvironment.java
@@ -1,0 +1,114 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.facebook.presto.hive;
+
+import com.facebook.presto.hive.authentication.NoHdfsAuthentication;
+import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.fs.FSDataInputStream;
+import org.apache.hadoop.fs.FSDataOutputStream;
+import org.apache.hadoop.fs.FileStatus;
+import org.apache.hadoop.fs.FileSystem;
+import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.fs.permission.FsPermission;
+import org.apache.hadoop.util.Progressable;
+
+import java.net.URI;
+
+public class TestingHdfsEnvironment
+        extends HdfsEnvironment
+{
+    public TestingHdfsEnvironment()
+    {
+        super(
+                new HiveHdfsConfiguration(new HdfsConfigurationUpdater(new HiveClientConfig(), new HiveS3Config())),
+                new HiveClientConfig(),
+                new NoHdfsAuthentication());
+    }
+
+    @Override
+    public FileSystem getFileSystem(String user, Path path, Configuration configuration)
+    {
+        return new TestingHdfsFileSystem();
+    }
+
+    private static class TestingHdfsFileSystem
+            extends FileSystem
+    {
+        @Override
+        public boolean delete(Path f, boolean recursive)
+        {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public boolean rename(Path src, Path dst)
+        {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public void setWorkingDirectory(Path dir)
+        {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public FileStatus[] listStatus(Path f)
+        {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public FSDataOutputStream create(Path f, FsPermission permission, boolean overwrite, int bufferSize, short replication, long blockSize, Progressable progress)
+        {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public boolean mkdirs(Path f, FsPermission permission)
+        {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public FSDataOutputStream append(Path f, int bufferSize, Progressable progress)
+        {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public FSDataInputStream open(Path f, int buffersize)
+        {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public FileStatus getFileStatus(Path f)
+        {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public Path getWorkingDirectory()
+        {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public URI getUri()
+        {
+            throw new UnsupportedOperationException();
+        }
+    }
+}
